// openapi/requests/types.gen.ts
export type Id = number;

export type UrlType = string;

export type User = {
  updatedAt: string;
  createdAt: string;
  teamId: string;
  image: UrlType | null;
  description: string;
  nickname: string;
  id: Id;
};

export type UserDetail = User & {
  mostFavoriteCategory: {
    name: string;
    id: Id;
  } | null;
  averageRating: number;
  reviewCount: number;
  followeesCount: number;
  followersCount: number;
  isFollowing: boolean;
};
// ...
export type MeResponse = UserDetail;

export type MeError = unknown;

export type UpdateMeData = {
  body: UpdateUserRequestBody;
  path: {
    teamId: string;
  };
};

export type UpdateMeResponse = User & {
  description: string;
};

export type UpdateMeError = unknown;